+++ packages/ws-controller/src/index.ts
@@ -12,6 +12,7 @@
 export { WebRtcTransportRequestorServer } from "@matter/node/behaviors/web-rtc-transport-requestor";
 export * from "./controller/ControllerCommandHandler.js";
 export * from "./controller/LegacyDataInjector.js";
+export * from "./controller/legacyStorageMigration.js";
 export * from "./controller/MatterController.js";
 export * from "./controller/OtaUploadRegistry.js";
 export * from "./controller/ServerIdResolver.js";